Remember client response
Upon connection, the server generates a random id to use as the key for the user's vote. It sends that key to the client as part of a handshake to determine if the client has already voted. The client looks for a cookie which is its ID, and if there is one, it disregards the new id, sending back the cookie. If there is no cookie, it just sends back the id provided by the server. The server then compares its id to the one sent back which identifies if the client is new or not. If it's new, it triggers the 'no vote' scenario which results in displaying ''You haven't voted yet'' on the page. If there is an existing vote, it displays a message confirming that existing vote, and when the user changes it, no new vote is created - instead, it just changes the vote for that client. This means it's more difficult for a single person to create multiple votes either by accident or mal intent.