Use post request for submission
Switched from using a socket to submit the poll to using a standard post request on the form instead.
Also implemented admins in preparation for viewing an admin page that's a varied version of the standard poll page...
Upon creation of a poll that uses a random string as its ID, I create a separate random string acting as the key in the admins object. The admin key's value is the polls ID. Now, when you access /admin/:id, it serves up the standard poll page with two small differences - instead of checking for the existence of polls[id], it checks for polls[admins[id]], and it passes 'admin: true' to the view.

The old client file with the old methodology is still in the file but will be removed.