Display access links
Added post-links.js which is called from poll.ejs in order display the public and admin links. This is necessary because you can't access the domain of the current URL from the .ejs file... instead, the script pulls the current domain after the document has loaded, merges it with the existing paths on the page, and replaces those paths with formatted html links.
Also modified server.js to send the necessary data to poll.ejs. Also added admin info to the demo data.