Implement voting
Setup proper socket client. The client is identified via cookie in order to allow revisiting of poll with memory of past vote. This is a better user experience and prevents accidental repeated votes.
Upon connection, the server initiates a handshake providing a new clientID. The client checks for an existing cooke - if none exists, he returns the ID provided by the server, but he returns the existing ID from the cookie if it does exist. Therefore, the server knows the client is new if it receives back the same ID that was sent, and the client is deemed old if the ID's are different.

Currently, there is one datastore for votes, and there is no differentiating between polls. This means the functionality will not work if the system includes more than one poll. This will be resolved in a later update.

Also, an admin cannot cast votes. The socket client currently identifies presence of an admin by checking the URL path. I don't love this solution, but it works unless I find a better solution. If it's an admin page, much of the socket client functionality doesn't happen, and the page calls a script to change the class of the buttons so they turn gray to indicate they aren't clickable.