Fix voting across polls
Now votes belong to specific polls.
Also changed from using cookies to localStorage which seems to work better. Before, it was bugging out and producing multiple cookies for a single user.
Also changed the format of the poll storage. Instead of a,b,c,d pointing to the text of the responses, the responsed themselves are now the keys and the values are all initialized to zero at creation. This sets things up to enable counting of votes and reporting the results.