Fix bug re: display of 'my vote' on private poll page
When loading a poll display page, the client displays the results on the page, but they're not supposed to if the poll is private. When the poll is private, there isn't even a div on the page to receive the results... so when trying to display them, they had nowhere to go and the script crashed. Now, the client checks if the poll is private before trying to load the results. This required sending the privacy setting to the client - accomplished this by sending over the whole poll instead of just the votes. Now it has the votes and privacy.
For that to work, I merged the votes object into the polls object so I don't have to pass around multiple objects.