Visualize results by changing button width
Removed the dedicated results area and make the buttons change width based on popularity, and the literal percentage is listed next to each button.
This special visualization only happens after at least one vote has been cast, and the width of the buttons can range from 50% - 100% (50% of the votes yields a button 75% wide).