Don't show blank buttons
Program allows up to four responses to be entered. Before, there were four buttons even if the poll creator submitted fewer responses. Now it only show buttons for which there is a possible response.
This was done by clearing out empty responses from the poll object and dynamically generating the response divs on the poll page.
Also needed to add a color-buttons.js script to add the bootstrap classes, like btn-danger and btn-primary, which is called after the response divs are created.