Change .admin-response class to .dead-button
In preparation for handling closed polls, I figured I'd use the same solution to close the poll as I already do with admin display (gray buttons and no .btn class which is what takes the event listener for the button press). That way, I'll just be able to call the kill-buttons.js script when it's time to close a poll.