Fix bug with .result-#id
In the poll view, it was iterating over the possible responses and applying this id dynamically - 'result-' + <%= response %> but it SHOULD be 'result-<%= response %>' (and is now).
Also, took out the default 'btn' class from each response and add that in the color-buttons.js script. This way, there's no chance of clicking on a button before the buttons have been adjusted to be active or closed based on other conditions.