Merge kill-butons.js and color-buttons.js
Merged kill-buttons and color-buttons into one file with logic to determine which function (killButtons or activateButtons) is called. This required setting a global variable to send in the state of 'admin', so I added <script var admin = <%= admin %></script> before calling the external script file. Adding this 'admin' global passed from the server means I no longer need the crappy noAdmin() function that was previously added to client.js (it just checked the URL to determine if it was an admin which isn't future proof).
This whole thing is a setup for only activating buttons if the poll hasn't closed.