Implement time-based closing of polls
Added logic in adjust-buttons.js to kill the buttons if the poll has expired and added the key 'expiration' to the poll object. Expiration is currently only set at the time of poll creation and uses bottstrap-datetimepicker to do it.
Still need to add messaging to state when a poll closes and that it is closed after the fact.
Also need the ability to change the expiration from the admin page.