Change expiration from admin page
Added the datetimepicker to the admin page and setup some socket action to make it work properly...
There's a 'Close poll' and 'Save new expiration' button. They're essentially the same except that the Close poll button sets the current time as the time to close, so whether it was an open poll or expiring (before or after current time), it's now closed with an expiration of now.
Also added messaging on the page to say when the poll closes (if it does). It states the time in 'calendar time' (like, 'tomorrow at 5:35 pm), but it just says 'Poll is closed.' if closed.
The input box on the admin page is set to the current closing time when the page loads, so hitting the 'Save new expiration' button doesn't accidentally remove the expiration if pressed.

Also discovered a huge bug I didn't find before - any server.sockets command affects all connections without regard to the poll the user is currently reviewing... for example, if you're looking at poll a and someone votes on poll b, your buttons will change size as if you were looking at poll b. I need to make the client only respond to messages relevant to its poll rather than all the polls.