Limit client reactions to only relevant clients
Before this fix, emissions to all clients weren't being filtered by the client, so votes on one poll would affect the results visualization on all other polls being viewed at the time. This also affected the display of poll closing time since that's emitted to all clients as well.
To resolve this, I store the client's confirmed poll ID, clientPollID, after the client completes the handshake with the server. I also send the affected pollID along with any fully broadcast message from the server. Then the client can confirm that the message is intended for it and ignore it if it is not.