Fix failure to activate buttons after poll opens from closed
Changed the pollExpired function to take an expiration instead of always using the global passed in from the view. This enables it to be called from the client with the new expiration added by the admin. It now automatically activates buttons and kills buttons as the poll's closed status changes. However, it does NOT automatically update if a user has the page loaded when the poll expires - that would require a refresh or an action from the admin to trigger the socket emission.