Add comments
Not finished, but the additional comments object has been added to the polls as have the necessary sockets. Currently, a user can add a comment, and it automatically appears on the page, but the comments are duplicating with any page load, so that needs to be fixed.
Also, identified a bug with polls accepting votes after closing if they closed while the user was viewing the page - simply put, the buttons gray out, but you can still click them. This is not a problem once the page is refreshed.