Fix private poll results display
Previously, the function visualizeResults was only checking if the poll was not private in order to show results. I added isAdminPage() to also render results if it on the admin view.

Also added in a total vote count on the results page so the admin knows how many have voted (though it's visible for everyone).